+++ zed_components/src/zed_camera/src/zed_camera_component_main.cpp
@@ -98,14 +98,14 @@ ZedCamera::ZedCamera(const rclcpp::NodeOptions & options)
   mUptimer(get_clock()),
   mSetSvoFrameCheckTimer(get_clock())
 {
+  mUsingIPC = options.use_intra_process_comms();
+
   RCLCPP_INFO(get_logger(), "================================");
   RCLCPP_INFO(get_logger(), "      ZED Camera Component ");
   RCLCPP_INFO(get_logger(), "================================");
   RCLCPP_INFO(get_logger(), " * namespace: %s", get_namespace());
   RCLCPP_INFO(get_logger(), " * node name: %s", get_name());
-  RCLCPP_INFO(
-    get_logger(), " * IPC: %s",
-    options.use_intra_process_comms() ? "enabled" : "disabled");
+  RCLCPP_INFO(get_logger(), " * IPC: %s", mUsingIPC ? "enabled" : "disabled");
   RCLCPP_INFO(get_logger(), "================================");
 
   if (((ZED_SDK_MAJOR_VERSION * 10 + ZED_SDK_MINOR_VERSION) <
@@ -649,12 +649,15 @@ void ZedCamera::getDebugParams()
     _debugPointCloud, _debugPointCloud,
     " * Debug Point Cloud: ");
   sl_tools::getParam(
-    shared_from_this(), "debug.debug_gnss", _debugGnss,
-    _debugGnss, " * Debug GNSS: ");
+    shared_from_this(), "debug.debug_tf", _debugTf, _debugTf,
+    " * Debug TF: ");
   sl_tools::getParam(
     shared_from_this(), "debug.debug_positional_tracking",
     _debugPosTracking, _debugPosTracking,
     " * Debug Positional Tracking: ");
+  sl_tools::getParam(
+    shared_from_this(), "debug.debug_gnss", _debugGnss,
+    _debugGnss, " * Debug GNSS: ");
   sl_tools::getParam(
     shared_from_this(), "debug.debug_sensors", _debugSensors,
     _debugSensors, " * Debug sensors: ");
@@ -682,9 +685,10 @@ void ZedCamera::getDebugParams()
     _debugAdvanced, " * Debug Advanced: ");
 
   mDebugMode = _debugCommon || _debugSim || _debugVideoDepth || _debugCamCtrl ||
-    _debugPointCloud || _debugPosTracking || _debugGnss ||
+    _debugPointCloud || _debugTf || _debugPosTracking || _debugGnss ||
     _debugSensors || _debugMapping || _debugObjectDet ||
-    _debugBodyTrk || _debugAdvanced || _debugRoi || _debugStreaming || _debugNitros;
+    _debugBodyTrk || _debugAdvanced || _debugRoi ||
+    _debugStreaming || _debugNitros;
 
   if (mDebugMode) {
     rcutils_ret_t res = rcutils_logging_set_logger_level(
@@ -1943,14 +1947,14 @@ rcl_interfaces::msg::SetParametersResult ZedCamera::callback_dynamicParamChange(
 void ZedCamera::setTFCoordFrameNames()
 {
   // ----> Coordinate frames
-  mCameraFrameId = mCameraName + "_camera_center";
+  mCenterFrameId = mCameraName + "_camera_center";
   mLeftCamFrameId = mCameraName + "_left_camera_frame";
   mLeftCamOptFrameId = mCameraName + "_left_camera_optical_frame";
   mRightCamFrameId = mCameraName + "_right_camera_frame";
   mRightCamOptFrameId = mCameraName + "_right_camera_optical_frame";
 
   mImuFrameId = mCameraName + "_imu_link";
-  mBaroFrameId = mCameraFrameId;         // mCameraName + "_baro_link";
+  mBaroFrameId = mCenterFrameId;         // mCameraName + "_baro_link";
   mMagFrameId = mImuFrameId;             // mCameraName + "_mag_link";
   mTempLeftFrameId = mLeftCamFrameId;    // mCameraName + "_temp_left_link";
   mTempRightFrameId = mRightCamFrameId;  // mCameraName + "_temp_right_link";
@@ -1964,7 +1968,7 @@ void ZedCamera::setTFCoordFrameNames()
   RCLCPP_INFO_STREAM(get_logger(), " * Map\t\t\t-> " << mMapFrameId);
   RCLCPP_INFO_STREAM(get_logger(), " * Odometry\t\t-> " << mOdomFrameId);
   RCLCPP_INFO_STREAM(get_logger(), " * Base\t\t-> " << mBaseFrameId);
-  RCLCPP_INFO_STREAM(get_logger(), " * Camera\t\t-> " << mCameraFrameId);
+  RCLCPP_INFO_STREAM(get_logger(), " * Camera\t\t-> " << mCenterFrameId);
   RCLCPP_INFO_STREAM(get_logger(), " * Left\t\t-> " << mLeftCamFrameId);
   RCLCPP_INFO_STREAM(
     get_logger(),
@@ -2297,12 +2301,16 @@ void ZedCamera::initPublishers()
     // ----> Camera/imu transform message
     if (mPublishSensImuTransf) {
       std::string cam_imu_tr_topic = mTopicRoot + "left_cam_imu_transform";
+      auto qos = mQos;
+      if (!mUsingIPC) {
+        // Use TRANSIENT_LOCAL durability if not using intra-process comms
+        qos = qos.durability(rclcpp::DurabilityPolicy::TransientLocal);
+      }
       mPubCamImuTransf = create_publisher<geometry_msgs::msg::TransformStamped>(
-        cam_imu_tr_topic, mQos, mPubOpt);
+        cam_imu_tr_topic, qos, mPubOpt);
 
       RCLCPP_INFO_STREAM(
-        get_logger(), " * Advertised on topic: "
-          << mPubCamImuTransf->get_topic_name());
+        get_logger(), " * Advertised on topic: " << mPubCamImuTransf->get_topic_name());
     }
 
     sl::Orientation sl_rot = mSlCamImuTransf.getOrientation();
@@ -2418,6 +2426,15 @@ bool ZedCamera::startCamera()
   mTfListener = std::make_unique<tf2_ros::TransformListener>(
     *mTfBuffer);    // Start TF Listener thread
   mTfBroadcaster = std::make_unique<tf2_ros::TransformBroadcaster>(this);
+
+  mStaticTfPublished = false;
+  mStaticImuTfPublished = false;
+  if (!mUsingIPC) {
+    mStaticTfBroadcaster =
+      std::make_unique<tf2_ros::StaticTransformBroadcaster>(this);
+  } else { // Cannot use LOCAL_TRANSIENT with intra-process comms
+    mStaticTfBroadcaster.reset();
+  }
   // <---- TF2 Transform
 
   // ----> ZED configuration
@@ -2445,13 +2462,14 @@ bool ZedCamera::startCamera()
   } else if (!mStreamAddr.empty()) {
     RCLCPP_INFO(get_logger(), "=== LOCAL STREAMING OPENING ===");
 
-    mInitParams.input.setFromStream(mStreamAddr.c_str(), static_cast<unsigned short>(mStreamPort));
+    mInitParams.input.setFromStream(
+      mStreamAddr.c_str(),
+      static_cast<unsigned short>(mStreamPort));
   } else {
     RCLCPP_INFO(get_logger(), "=== CAMERA OPENING ===");
 
     mInitParams.camera_fps = mCamGrabFrameRate;
-    //mInitParams.grab_compute_capping_fps = static_cast<float>(mVdPubRate); // Using Wrapper multi-threading instead
-    mInitParams.grab_compute_capping_fps = 0.0f;
+    mInitParams.grab_compute_capping_fps = 0.0f; // Using multi-threading in the component
     mInitParams.camera_resolution = static_cast<sl::RESOLUTION>(mCamResol);
     mInitParams.async_image_retrieval = mAsyncImageRetrieval;
     mInitParams.enable_image_validity_check = mImageValidityCheck;
@@ -2643,8 +2661,9 @@ bool ZedCamera::startCamera()
       mVdPubRate = mCamGrabFrameRate;
       RCLCPP_WARN_STREAM(
         get_logger(),
-        "Video/Depth publishing rate was too high [" << mVdPubRate << "], capped to real grab rate: " <<
-          mCamGrabFrameRate);
+        "Video/Depth publishing rate was too high ["
+          << mVdPubRate << "], capped to real grab rate: "
+          << mCamGrabFrameRate);
     }
     if (mPcPubRate > mCamGrabFrameRate) {
       mPcPubRate = mCamGrabFrameRate;
@@ -3894,8 +3913,8 @@ void ZedCamera::initTransforms()
 
 bool ZedCamera::getCamera2BaseTransform()
 {
-  DEBUG_STREAM_PT(
-    "Getting static TF from '" << mCameraFrameId.c_str()
+  DEBUG_STREAM_TF(
+    "Getting static TF from '" << mCenterFrameId.c_str()
                                << "' to '" << mBaseFrameId.c_str()
                                << "'");
 
@@ -3906,7 +3925,7 @@ bool ZedCamera::getCamera2BaseTransform()
   try {
     // Save the transformation
     geometry_msgs::msg::TransformStamped c2b = mTfBuffer->lookupTransform(
-      mCameraFrameId, mBaseFrameId, TIMEZERO_SYS, rclcpp::Duration(1, 0));
+      mCenterFrameId, mBaseFrameId, TIMEZERO_SYS, rclcpp::Duration(1, 0));
 
     // Get the TF2 transformation
     // tf2::fromMsg(c2b.transform, mCamera2BaseTransf);
@@ -3924,7 +3943,7 @@ bool ZedCamera::getCamera2BaseTransform()
     RCLCPP_INFO(
       get_logger(),
       " Static transform Camera Center to Base [%s -> %s]",
-      mCameraFrameId.c_str(), mBaseFrameId.c_str());
+      mCenterFrameId.c_str(), mBaseFrameId.c_str());
     RCLCPP_INFO(
       get_logger(), "  * Translation: {%.3f,%.3f,%.3f}",
       mCamera2BaseTransf.getOrigin().x(),
@@ -3942,7 +3961,7 @@ bool ZedCamera::getCamera2BaseTransform()
       RCLCPP_WARN_THROTTLE(
         get_logger(), steady_clock, 1000.0,
         "The tf from '%s' to '%s' is not available.",
-        mCameraFrameId.c_str(), mBaseFrameId.c_str());
+        mCenterFrameId.c_str(), mBaseFrameId.c_str());
       RCLCPP_WARN_THROTTLE(
         get_logger(), steady_clock, 1000.0,
         "Note: one of the possible cause of the problem is the absense of an "
@@ -3951,7 +3970,7 @@ bool ZedCamera::getCamera2BaseTransform()
         "TF transformations "
         "or a modified URDF not correctly reproducing the ZED "
         "TF chain '%s' -> '%s' -> '%s'",
-        mBaseFrameId.c_str(), mCameraFrameId.c_str(), mDepthFrameId.c_str());
+        mBaseFrameId.c_str(), mCenterFrameId.c_str(), mDepthFrameId.c_str());
       mCamera2BaseFirstErr = false;
     }
 
@@ -3966,9 +3985,9 @@ bool ZedCamera::getCamera2BaseTransform()
 
 bool ZedCamera::getSens2CameraTransform()
 {
-  DEBUG_STREAM_PT(
+  DEBUG_STREAM_TF(
     "Getting static TF from '"
-      << mDepthFrameId.c_str() << "' to '" << mCameraFrameId.c_str()
+      << mDepthFrameId.c_str() << "' to '" << mCenterFrameId.c_str()
       << "'");
 
   mSensor2CameraTransfValid = false;
@@ -3978,7 +3997,7 @@ bool ZedCamera::getSens2CameraTransform()
   try {
     // Save the transformation
     geometry_msgs::msg::TransformStamped s2c = mTfBuffer->lookupTransform(
-      mDepthFrameId, mCameraFrameId, TIMEZERO_SYS, rclcpp::Duration(1, 0));
+      mDepthFrameId, mCenterFrameId, TIMEZERO_SYS, rclcpp::Duration(1, 0));
 
     // Get the TF2 transformation
     // tf2::fromMsg(s2c.transform, mSensor2CameraTransf);
@@ -3996,7 +4015,7 @@ bool ZedCamera::getSens2CameraTransform()
     RCLCPP_INFO(
       get_logger(),
       " Static transform ref. CMOS Sensor to Camera Center [%s -> %s]",
-      mDepthFrameId.c_str(), mCameraFrameId.c_str());
+      mDepthFrameId.c_str(), mCenterFrameId.c_str());
     RCLCPP_INFO(
       get_logger(), "  * Translation: {%.3f,%.3f,%.3f}",
       mSensor2CameraTransf.getOrigin().x(),
@@ -4014,7 +4033,7 @@ bool ZedCamera::getSens2CameraTransform()
       RCLCPP_WARN_THROTTLE(
         get_logger(), steady_clock, 1000.0,
         "The tf from '%s' to '%s' is not available.",
-        mDepthFrameId.c_str(), mCameraFrameId.c_str());
+        mDepthFrameId.c_str(), mCenterFrameId.c_str());
       RCLCPP_WARN_THROTTLE(
         get_logger(), steady_clock, 1000.0,
         "Note: one of the possible cause of the problem is the absense of an "
@@ -4023,7 +4042,7 @@ bool ZedCamera::getSens2CameraTransform()
         "TF transformations "
         "or a modified URDF not correctly reproducing the ZED "
         "TF chain '%s' -> '%s' -> '%s'",
-        mBaseFrameId.c_str(), mCameraFrameId.c_str(), mDepthFrameId.c_str());
+        mBaseFrameId.c_str(), mCenterFrameId.c_str(), mDepthFrameId.c_str());
       mSensor2CameraTransfFirstErr = false;
     }
 
@@ -4038,7 +4057,7 @@ bool ZedCamera::getSens2CameraTransform()
 
 bool ZedCamera::getSens2BaseTransform()
 {
-  DEBUG_STREAM_PT(
+  DEBUG_STREAM_TF(
     "Getting static TF from '" << mDepthFrameId.c_str()
                                << "' to '" << mBaseFrameId.c_str()
                                << "'");
@@ -4095,7 +4114,7 @@ bool ZedCamera::getSens2BaseTransform()
         "TF transformations "
         "or a modified URDF not correctly reproducing the ZED "
         "TF chain '%s' -> '%s' -> '%s'",
-        mBaseFrameId.c_str(), mCameraFrameId.c_str(), mDepthFrameId.c_str());
+        mBaseFrameId.c_str(), mCenterFrameId.c_str(), mDepthFrameId.c_str());
       mSensor2BaseTransfFirstErr = false;
     }
 
@@ -4229,6 +4248,15 @@ bool ZedCamera::setPose(
 
 void ZedCamera::publishImuFrameAndTopic()
 {
+  if (!mPublishSensImuTransf && !mPublishImuTF) {
+    return;
+  }
+
+  if (!mUsingIPC && mStaticImuTfPublished) {
+    DEBUG_ONCE_TF("Static Imu TF and Transient Local message already published");
+    return;
+  }
+
   sl::Orientation sl_rot = mSlCamImuTransf.getOrientation();
   sl::Translation sl_tr = mSlCamImuTransf.getTranslation();
 
@@ -4249,7 +4277,9 @@ void ZedCamera::publishImuFrameAndTopic()
   cameraImuTransfMgs->transform.translation.z = sl_tr.z;
 
   try {
-    if (mPubCamImuTransf) {mPubCamImuTransf->publish(std::move(cameraImuTransfMgs));}
+    if (mPubCamImuTransf) {
+      mPubCamImuTransf->publish(std::move(cameraImuTransfMgs));
+    }
   } catch (std::system_error & e) {
     DEBUG_STREAM_COMM("Message publishing exception: " << e.what());
   } catch (...) {
@@ -4281,13 +4311,47 @@ void ZedCamera::publishImuFrameAndTopic()
   transformStamped.transform.translation.y = sl_tr.y;
   transformStamped.transform.translation.z = sl_tr.z;
 
-  mTfBroadcaster->sendTransform(transformStamped);
+  if (mUsingIPC) {
+    mTfBroadcaster->sendTransform(transformStamped);
+    DEBUG_STREAM_TF(
+      "Broadcasted new IMU dynamic transform: "
+        << transformStamped.header.frame_id << " -> " << transformStamped.child_frame_id);
+  } else {
+    mStaticTfBroadcaster->sendTransform(transformStamped);
+    DEBUG_STREAM_TF(
+      "Broadcasted new IMU static transform: "
+        << transformStamped.header.frame_id << " -> " << transformStamped.child_frame_id);
+  }
   // <---- Publish TF
 
   // IMU TF publishing diagnostic
   double elapsed_sec = mImuTfFreqTimer.toc();
   mPubImuTF_sec->addValue(elapsed_sec);
   mImuTfFreqTimer.tic();
+
+  // Debug info
+  if (_debugTf) {
+    double roll, pitch, yaw;
+    tf2::Matrix3x3(
+      tf2::Quaternion(
+        transformStamped.transform.rotation.x,
+        transformStamped.transform.rotation.y,
+        transformStamped.transform.rotation.z,
+        transformStamped.transform.rotation.w))
+    .getRPY(roll, pitch, yaw);
+    DEBUG_STREAM_TF(
+      "TF ["
+        << transformStamped.header.frame_id << " -> "
+        << transformStamped.child_frame_id << "] Position: ("
+        << transformStamped.transform.translation.x << ", "
+        << transformStamped.transform.translation.y << ", "
+        << transformStamped.transform.translation.z
+        << ") - Orientation RPY: (" << roll * RAD2DEG << ", "
+        << pitch * RAD2DEG << ", " << yaw * RAD2DEG << ")");
+  }
+
+  // At the end
+  mStaticImuTfPublished = true;
 }
 
 void ZedCamera::threadFunc_zedGrab()
@@ -5137,17 +5201,19 @@ bool ZedCamera::publishSensorsData(rclcpp::Time force_ts)
 
 void ZedCamera::publishTFs(rclcpp::Time t)
 {
-  // DEBUG_STREAM_PT("publishTFs");
+  // DEBUG_STREAM_TF("publishTFs");
 
   // RCLCPP_INFO_STREAM(get_logger(), "publishTFs - t type:" <<
   // t.get_clock_type());
+
+  publishCameraTFs(t);
 
   if (!mPosTrackingReady) {
     return;
   }
 
   if (t == TIMEZERO_ROS) {
-    DEBUG_STREAM_PT("Time zero: not publishing TFs");
+    DEBUG_STREAM_TF("Time zero: not publishing TFs");
     return;
   }
 
@@ -5161,9 +5227,186 @@ void ZedCamera::publishTFs(rclcpp::Time t)
   }
 }
 
+void ZedCamera::publishCameraTFs(rclcpp::Time t)
+{
+  // DEBUG_STREAM_TF("publishCameraTFs");
+
+  if (!mUsingIPC && mStaticTfPublished) {
+    DEBUG_ONCE_TF("Static Camera TF already broadcasted");
+    return;
+  }
+
+  auto calib_params = mZed->getCameraInformation().camera_configuration.calibration_parameters;
+  const double baseline = static_cast<double>(calib_params.getCameraBaseline());
+  sl::Transform stereo_transform = calib_params.stereo_transform;
+
+  DEBUG_STREAM_TF("Calibrated Camera baseline: " << baseline << " m");
+  DEBUG_STREAM_TF(
+    "SDK Stereo Transform T: [" << stereo_transform.getTranslation().x << ", "
+                                << stereo_transform.getTranslation().y << ", "
+                                << stereo_transform.getTranslation().z << "]");
+  DEBUG_STREAM_TF(
+    "SDK Stereo Transform R: ["
+      << stereo_transform.getOrientation().x << ", "
+      << stereo_transform.getOrientation().y << ", "
+      << stereo_transform.getOrientation().z << ", "
+      << stereo_transform.getOrientation().w << "]");
+
+  // ----> Validate data
+  bool not_valid = false;
+  if (stereo_transform.getTranslation().x != 0.0 ||
+    stereo_transform.getTranslation().z != 0.0)
+  {
+    RCLCPP_WARN_STREAM(
+      get_logger(),
+      "Unexpected calibrated stereo transform translation: ["
+        << stereo_transform.getTranslation().x << ", "
+        << stereo_transform.getTranslation().y << ", "
+        << stereo_transform.getTranslation().z << "]. Expected [0, " << baseline << ", 0].");
+    not_valid = true;
+  }
+  if (stereo_transform.getOrientation().x != 0.0 ||
+    stereo_transform.getOrientation().y != 0.0 ||
+    stereo_transform.getOrientation().z != 0.0 ||
+    stereo_transform.getOrientation().w != 1.0)
+  {
+    RCLCPP_WARN_STREAM(
+      get_logger(),
+      "Unexpected calibrated stereo transform rotation: ["
+        << stereo_transform.getOrientation().x << ", "
+        << stereo_transform.getOrientation().y << ", "
+        << stereo_transform.getOrientation().z << ", "
+        << stereo_transform.getOrientation().w << "]. Expected [0, 0, 0, 1].");
+    not_valid = true;
+  }
+  if (baseline != -stereo_transform.getTranslation().y) {
+    RCLCPP_WARN_STREAM(
+      get_logger(),
+      "Baseline mismatch: Camera baseline is " << baseline << " m but calibrated stereo transform y-translation is " <<
+        stereo_transform.getTranslation().y << " m.");
+    not_valid = true;
+  }
+
+  if (not_valid) {
+    RCLCPP_WARN_STREAM(
+      get_logger(), "Please report this problem to Stereolabs support if you see this message, "
+      "adding information about your camera model and serial number: "
+        << sl::toString(mCamRealModel) << " - " << mCamSerialNumber);
+    return;
+  }
+  // <---- Validate data
+
+  double optical_offset_x = 0.0;
+
+  switch (mCamRealModel) {
+    case sl::MODEL::ZED:
+      optical_offset_x = -0.01;
+      break;
+    case sl::MODEL::ZED_M:
+      optical_offset_x = 0.0;
+      break;
+    case sl::MODEL::ZED2:
+      optical_offset_x = -0.01;
+      break;
+    case sl::MODEL::ZED2i:
+      optical_offset_x = -0.01;
+      break;
+    case sl::MODEL::ZED_X:
+    case sl::MODEL::ZED_XM:
+    case sl::MODEL::ZED_X_HDR:
+    case sl::MODEL::ZED_X_HDR_MAX:
+    case sl::MODEL::ZED_X_HDR_MINI:
+      optical_offset_x = -0.01;
+      break;
+    case sl::MODEL::VIRTUAL_ZED_X:
+      optical_offset_x = -0.01;
+      break;
+    default:
+      RCLCPP_ERROR_STREAM(
+        get_logger(),
+        "Unknown camera model for static TF publishing: "
+          << sl::toString(mCamRealModel).c_str());
+      mZed.reset();
+      exit(EXIT_FAILURE);
+  }
+
+  // ----> Lambda function to publish transform with debug info
+  auto publishTransform = [&](const geometry_msgs::msg::TransformStamped & tf) {
+      if (mUsingIPC) {
+        mTfBroadcaster->sendTransform(tf);
+        DEBUG_STREAM_TF(
+          "Broadcasted new dynamic transform: "
+            << tf.header.frame_id << " -> " << tf.child_frame_id);
+      } else {
+        mStaticTfBroadcaster->sendTransform(tf);
+        DEBUG_STREAM_TF(
+          "Broadcasted new static transform: "
+            << tf.header.frame_id << " -> " << tf.child_frame_id);
+      }
+      if (!_debugTf) {
+        double roll, pitch, yaw;
+        tf2::Matrix3x3(
+          tf2::Quaternion(
+            tf.transform.rotation.x, tf.transform.rotation.y,
+            tf.transform.rotation.z, tf.transform.rotation.w)).getRPY(roll, pitch, yaw);
+        DEBUG_STREAM_TF(
+          "TF ["
+            << tf.header.frame_id << " -> " << tf.child_frame_id
+            << "] Position: (" << tf.transform.translation.x << ", "
+            << tf.transform.translation.y << ", "
+            << tf.transform.translation.z << ") - Orientation RPY: ("
+            << roll * RAD2DEG << ", " << pitch * RAD2DEG << ", "
+            << yaw * RAD2DEG << ")");
+      }
+    };
+  // <---- Lambda function to publish transform with debug info
+
+  // ----> Common info
+  geometry_msgs::msg::TransformStamped transformStamped;
+  transformStamped.header.stamp =
+    mUsePubTimestamps ? get_clock()->now() :
+    (t + rclcpp::Duration(0, mTfOffset * 1e9));
+  // <---- Common info
+
+  // ----> New camera_center -> left_camera_frame
+  transformStamped.header.frame_id = mCenterFrameId;
+  transformStamped.child_frame_id = mLeftCamFrameId;
+
+  transformStamped.transform.translation.x = optical_offset_x;
+  transformStamped.transform.translation.y = baseline / 2.0;
+  transformStamped.transform.translation.z = 0.0;
+  transformStamped.transform.rotation.x = 0.0;
+  transformStamped.transform.rotation.y = 0.0;
+  transformStamped.transform.rotation.z = 0.0;
+  transformStamped.transform.rotation.w = 1.0;
+
+  // Publish transformation
+  publishTransform(transformStamped);
+  // <---- New camera_center -> left_camera_frame
+
+  // ----> New camera_center -> right_camera_frame
+  transformStamped.header.frame_id = mCenterFrameId;
+  transformStamped.child_frame_id = mRightCamFrameId;
+
+  transformStamped.transform.translation.x = optical_offset_x;
+  transformStamped.transform.translation.y = -baseline / 2.0;
+  transformStamped.transform.translation.z = 0.0;
+  transformStamped.transform.rotation.x = 0.0;
+  transformStamped.transform.rotation.y = 0.0;
+  transformStamped.transform.rotation.z = 0.0;
+  transformStamped.transform.rotation.w = 1.0;
+
+  // Publish transformation
+  publishTransform(transformStamped);
+  // <---- New camera_center -> right_camera_frame
+
+  // at the end
+  mStaticTfPublished = true;
+}
+
 void ZedCamera::publishOdomTF(rclcpp::Time t)
 {
-  // DEBUG_STREAM_PT("publishOdomTF");
+  // DEBUG_STREAM_TF("publishOdomTF");
 
   // ----> Avoid duplicated TF publishing
   if (t == mLastTs_odom) {
@@ -5213,11 +5456,31 @@ void ZedCamera::publishOdomTF(rclcpp::Time t)
   double elapsed_sec = mOdomFreqTimer.toc();
   mPubOdomTF_sec->addValue(elapsed_sec);
   mOdomFreqTimer.tic();
+
+  // Debug info
+  if (_debugTf) {
+    double roll, pitch, yaw;
+    tf2::Matrix3x3(
+      tf2::Quaternion(
+        transformStamped.transform.rotation.x,
+        transformStamped.transform.rotation.y,
+        transformStamped.transform.rotation.z,
+        transformStamped.transform.rotation.w))
+    .getRPY(roll, pitch, yaw);
+    DEBUG_STREAM_TF(
+      "TF [" << transformStamped.header.frame_id << " -> "
+             << transformStamped.child_frame_id << "] Position: ("
+             << transformStamped.transform.translation.x << ", "
+             << transformStamped.transform.translation.y << ", "
+             << transformStamped.transform.translation.z
+             << ") - Orientation RPY: (" << roll * RAD2DEG << ", "
+             << pitch * RAD2DEG << ", " << yaw * RAD2DEG << ")");
+  }
 }
 
 void ZedCamera::publishPoseTF(rclcpp::Time t)
 {
-  // DEBUG_STREAM_PT("publishPoseTF");
+  // DEBUG_STREAM_TF("publishPoseTF");
 
   // ----> Avoid duplicated TF publishing
   if (t == mLastTs_pose) {
@@ -5263,6 +5526,26 @@ void ZedCamera::publishPoseTF(rclcpp::Time t)
   double elapsed_sec = mPoseFreqTimer.toc();
   mPubPoseTF_sec->addValue(elapsed_sec);
   mPoseFreqTimer.tic();
+
+  // Debug info
+  if (_debugTf) {
+    double roll, pitch, yaw;
+    tf2::Matrix3x3(
+      tf2::Quaternion(
+        transformStamped.transform.rotation.x,
+        transformStamped.transform.rotation.y,
+        transformStamped.transform.rotation.z,
+        transformStamped.transform.rotation.w))
+    .getRPY(roll, pitch, yaw);
+    DEBUG_STREAM_TF(
+      "TF [" << transformStamped.header.frame_id << " -> "
+             << transformStamped.child_frame_id << "] Position: ("
+             << transformStamped.transform.translation.x << ", "
+             << transformStamped.transform.translation.y << ", "
+             << transformStamped.transform.translation.z
+             << ") - Orientation RPY: (" << roll * RAD2DEG << ", "
+             << pitch * RAD2DEG << ", " << yaw * RAD2DEG << ")");
+  }
 }
 
 void ZedCamera::threadFunc_pubSensorsData()
@@ -6186,6 +6469,27 @@ void ZedCamera::processGeoPose()
       (tf2::toMsg(mMap2UtmTransf.inverse()));
 
     mTfBroadcaster->sendTransform(transformStamped);
+
+    // Debug info
+    if (_debugTf) {
+      double roll, pitch, yaw;
+      tf2::Matrix3x3(
+        tf2::Quaternion(
+          transformStamped.transform.rotation.x,
+          transformStamped.transform.rotation.y,
+          transformStamped.transform.rotation.z,
+          transformStamped.transform.rotation.w))
+      .getRPY(roll, pitch, yaw);
+      DEBUG_STREAM_TF(
+        "TF ["
+          << transformStamped.header.frame_id << " -> "
+          << transformStamped.child_frame_id << "] Position: ("
+          << transformStamped.transform.translation.x << ", "
+          << transformStamped.transform.translation.y << ", "
+          << transformStamped.transform.translation.z
+          << ") - Orientation RPY: (" << roll * RAD2DEG << ", "
+          << pitch * RAD2DEG << ", " << yaw * RAD2DEG << ")");
+    }
   }
 
   publishGnssPose();
@@ -7469,6 +7773,7 @@ void ZedCamera::callback_updateDiagnostic(
   stat.addf("Uptime", "%s", sl_tools::seconds2str(mUptimer.toc()).c_str());
 
   if (mGrabStatus == sl::ERROR_CODE::SUCCESS || mGrabStatus == sl::ERROR_CODE::CORRUPTED_FRAME) {
+    stat.addf("IPC Enabled", "%s", mUsingIPC ? "YES" : "NO");
     stat.addf("Camera Grab rate", "%d Hz", mCamGrabFrameRate);
 
     double freq = 1. / mGrabPeriodMean_sec->getAvg();