+++ CHANGELOG.rst
@@ -1,6 +1,16 @@
 LATEST CHANGES
 ==============
 
+2025-11-09
+----------
+- Added debug option for TF broadcasting
+  - Improved TF debug logs to show frame transformations when enabled 
+- Static baseline information from URDF is now overwritten by the real baseline value retrieved from the camera calibration file.
+- Removed mandatory `custom_baseline` launch argument for virtual stereo cameras made with two ZED X One cameras. 
+  The value is retrieved from the calibration file.
+- IMU TF is not broadcasted as static if IPC is disabled.
+- IMU Transform topic is now published with TRANSIENT LOCAL durability if IPC is disabled.
+
 2025-11-05
 ----------
 - Changed minimum depth value to 0.01 meters when using ZED SDK v5.1 or higher